requirements.txt
suanpan-core>=0.16.8
flask>=2.0.2